Built preliminary manifestation system
The game is now able to go to and from manifestation states. The game keeps track of what type of tower is being manifested, changes the game's controls so Z exits manifestation rather than opening a menu (and X calls functions depending on the manifestation). The coordinates of the tower being manifested are saved and there's a slot for defining special logic at the time of manifestation and unmanifestation based on the tower.
Using this, I'm going to properly make the manifestation mode for the Hale Howitzer.