Finished Spyplane and Carrier
Implemented and merged the spyplane and carrier. The spyplane uses RNG to stealth and destealth, and while it can look a little flicker it makes them legitimately feel a little threatening (though not frustrating luckily, as towers will fire on them as normal. It'd be a good thing to get feedback on nonetheless). The Carrier literally just morphs into a car to save token count, though in gameplay it works surprisingly well (it includes the special weakness to the sharpshooter).