Improved tutorial to describe specific buttons and where to place the tower
A big issue I saw with playtesting in person is that people, in the ifrst time playing, will immediately throw down their sword circle in the middle of the map and have to restart. I made some changes (including a marker for where to place it and better tutorial dialogue) to help avoid this starting frustration. Also got rid of the unused tileset again, and rebalanced the sword circle (it felt a bit strong with the minor buff to defeat cars, so I just lowered car HP from 12 to 11).